+++ pages/api/trpc/[trpc].ts
@@ -1,5 +1,5 @@
 import { createNextApiHandler } from "@trpc/server/adapters/next";
-import { appRouter } from "../../../server/api/routers/_app";
+import { appRouter } from "../../../server/api/root";
 import { createTRPCContext } from "../../../server/api/trpc";
 
 export default createNextApiHandler({
