+++ server/api/root.ts
@@ -1,0 +1,13 @@
+import { createTRPCRouter } from "./trpc";
+import { helloRouter } from "./routers/hello";
+
+/**
+ * This is the primary router for your server.
+ *
+ * All routers added in /api/routers should be manually added here.
+ */
+export const appRouter = createTRPCRouter({
+  hello: helloRouter,
+});
+
+export type AppRouter = typeof appRouter;
