+++ utils/api.ts
@@ -2,7 +2,7 @@ import { httpBatchLink, loggerLink } from "@trpc/client";
 import { createTRPCNext } from "@trpc/next";
 import superjson from "superjson";
 
-import { type AppRouter } from "../server/api/routers/_app";
+import { type AppRouter } from "../server/api/root";
 import { type inferRouterInputs, type inferRouterOutputs } from "@trpc/server";
 
 function getBaseUrl() {
